backend/requirements.txt
bs4
pyaudio
requests
decorator
aiohttp
aiohttp-cors
python-socketio
pywin32
easygui